Migrate run_gp_tool to shared helpers module
